# homefwork5/go20240606/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample in go tutorial for building a short URL service",
        "title": "Simple Short URL API Server",
        "contact": {
            "name": "Leon Lin",
            "url": "github.com/leon123858",
            "email": "a0970785699@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
        "/admin/register": {
            "post": {
                "description": "Create a new admin user with the provided email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Create admin user",
                "parameters": [
                    {
                        "description": "Admin details",
                        "name": "admin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.CreateAdminRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created admin user",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/admin/statistic/{password}": {
            "get": {
                "description": "Get statistics for an admin user with the provided password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get admin statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Admin password",
                        "name": "password",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statistics events",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/pg.Event"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/shorten": {
            "post": {
                "description": "Create a new short URL for a given long URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URLs"
                ],
                "summary": "Create short URL",
                "parameters": [
                    {
                        "description": "URL to be shortened",
                        "name": "url",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/url.ShortenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Short URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/{token}": {
            "get": {
                "description": "Get the long URL for a given short URL token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URLs"
                ],
                "summary": "Get long URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short URL token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to long URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "URL not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "admin.CreateAdminRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "pg.Event": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "url.ShortenRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}