# homefwork5/go20240606/pkg/pg/model.go
package pg

type Event struct {
	Password string `json:"password" gorm:"foreignKey:users(password)"`
	Time     string `json:"time"`
	Url      string `json:"url"`
}

type User struct {
	Email    string `json:"email"`
	Password string `json:"password" gorm:"primaryKey"`
}